+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
 ### Implemented enhancements
 - TT-2322 - Route /local-export-support/apply/ emails to closest trade office
+- TT-2323 - Add ecommerce export support form
 
 ### Bugs fixed
 - No ticket - Upgrade markdown to fix vulnerability
+- No ticket - Fixed local export support form
 
 ## [2020.05.18](https://github.com/uktrade/great-domestic-ui/releases/tag/2020.05.18_1)
 [Full Changelog](https://github.com/uktrade/great-domestic-ui/compare/2020.04.22...2020.05.18_1)